+++ watchtower-app/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2024-08-27
+### Added
+- Add support for null (`-`) values in CSV files
+- Mark some attributes values as null when the device also has a "switch" attribute with value "off"
+- Add attributes support: Alarm/alarm, CarbonMonoxideDetector/carbonMonoxide, Chime/status, AudioVolume/volume, SwitchLevel/level, UltravioletIndex/ultravioletIndex
+
 ## [1.4.1] - 2024-08-25
 ### Fixed
 - Fix "Max records with 5 min accuracy" settings - `@bobbles`